requirements.txt
Django>=4.2
django-jazzmin
Pillow
psycopg2-binary
python-decouple
djangorestframework
